Create LICENSE
Determined that this will be maintained and published with the GNU General Public LIcense, v3.

This is not retroactive to previous commits or to code I did not write which was licensed to me but could not be relicensed under these terms and is freely licensed and available elsewhere but not critical to this project. While the majority of this code is mine, some compatible code was removed because it no longer fit with the goals. The dynamic inventory code which remained after my edits were under a compatible license that permits relicensing under the GPL with its attribution still intact.

Prior to me licensing this code, there were no terms that unambiguously permitted downloading, copying, modifying, or distributing. That does not make that code any more or less free, it is the license here that extends those rights to the current and future versions to others.