Templated overrides. Fixed firewall for WebRTC.
Overall the templating was a success. 9 lines can fully generate override configuration for both the TESmart and ezCoo KVM switches, both protocols of ezcoo and supporting the 2, 4, 8, and 16-port models using a number range. The firewall was missing the UDP port range for Janus and WebRTC, I noticed this while testing the switch and added the port range. I stubbed out the firewall config in the group_vars for all hosts and updated the comments there, the public.xml template currently doesn't use this yet.

Note, at present if any of the variables are undefined for the overrides that aren't checked, the overrides won't be generated and the recovery behavior is to unnecessarily reboot PiKVM. Worse, the registered handlers will reboot them again if the packages haven't been updated. The same reboot occured if the firewall rules were changed because of a variable change, that has been fixed.