Replaced port_range with size and added defaults
Some of the devices represented by specific drivers do have vendor or acceptable defaults, such as TESMart's host being set to 192.168.1.10, the port being set to 5000. The ezCoo defaults come from default Linux (udev rules) for the /dev/ttyUSB0 default device and PiKVM meanings associated with protocols 1 and 2.

If a default is used, a comment "# Default for <device>" will be added after the setting.

This is everything I needed to complete my core configurations. There's been some cleanup of the templates to make them more usable, some comments added to code, I've updated two of my host configurations, pikvm1 and pikvm3, and made it easier to run the test playbook that generates test configurations in the local /tmp/ directory based on the inventory_hostname.