Physically tested the old 2-port ezCoo and changed "hosts" to "host_list"
The bookshelf lab was originally a 2-port ezCoo USB 3.0 switch and this lab never had more than 2 machines. That said, I had swapped out the 2-port switch with the 4-port switch to make sure it arrived in working order and never reconfigured it back.

The extra variable "hosts" is also an Ansible reserved word so, changing the override for the host list to "host_list" made sense. Rather than switching all devices to read-write for a long playbook execution, I'm being safer and faster updating only the host(s) I need to update and leaving the rest untouched.

Especially if I'm not sure my code is working. I accidentally scrambled the root password on all of the hosts. This gets saved to a file but I'm using SSH public key authentication so all I had to do was put the right password in the files and re-run the playbook. That said, I would have rather run that on one host rather than all 7 defined hosts, so I spent less time waiting for the playbook to fix everything.

It's possible to have some tasks run only when certain tags are specified. I'll probably implement that when I get some spare time, Ansible tends to be a bit slow. That's fine when it's running in the background, not so good when I haven't touched a task, I don't need to run it again to test it, but I'm waiting on it to complete anyway to verify no changes are needed.