Transition playbook to role and fix task blocking
Major set of changes. This has changed quite a bit from a simple playbook to configure PiKVM into more of a role with tasks and task lists, blocking, templates, etc. Overall it would have made sense to start this as a role to begin with rather than as just a playbook. The blocking has been fixed and the ansible_lint errors have all been addressed.

More importantly, the playbook still appears to work the way it should!