Tags added to save time and wear on your PiKVM!
It was taking a long time to test because everytime I ran, even with a single host, it ran every single configuration and security baseline check. Check if passwords are set or need to be changed, render overrides and see if they match, etc.

Now, when it's called with `ansible-playbook main.yml --host_list=pikvm3 --tags=override_config`, it will only render the template and write the configuration if it doesn't match. Note, the SD card still needs to be read-write, it's not doing two passes, but it no longer checks for updates, verifies firewall configuration, etc.

To see the available tags:

```
$ ansible-playbook --list-tags main.yml

playbook: main.yml

  play #1 (all): Playbook to configure PiKVM Hosts      TAGS: []
      TASK TAGS: [baseline_config, override_config, package_maintenance, security_config]

  play #2 (localhost): Print options for the user       TAGS: []
      TASK TAGS: []
[bbeaudoin@cerebus ansible-pikvm]$ ansible-playbook --tags package_maintenance main.yml
```

If the playbook is called with no tags, there will be a debug message shown when tags are set to "all". Because all tasks include the "never" tag, all isn't exactly an error, it just doesn't match anything but the task that tries to be helpful.