Added support multi-level KVM switches. Renamed role and tags.
The major change here is multi-level KVM switches is supported. I had planned on using "switch0_button0" but was troubleshooting the table generation, turns out it was an issue with the spacing in the Jinja template. How to use the index to get the current loop iteration had multiple conflicting sources but found a good source, considering 1-based indexing for readability.

Conceptually having two 4-port switches behind a 2-port switch isn't technically hard, if there are enough ports and one knows which USB device index for the device, one can switch all three independently. In theory, the 256 limit of TESmart KVM switches could be easily reached easily with this configuration with 17 switch definitions (16*16 ports) though the UI might not be able to really scale to that. However, having a 4-port ezCoo with multiple 4-port ezCoo switches looks like it would fit just fine.