More declarative, better tags, improved firewall and package handling.
Sudo rules should only be added when needed and defined so, this is now a list that should be moved to the inventory for hosts that need to add extra permissions to use things like a "Reboot" or "Restart kvmd" option to the UI. The firewall configuration now depends on both a configuration being defined and the package in the package_list.

Speaking of the package_list, it may be necessary for a host to either extend or replace the group_packages so both now use a single definition with the host_vars overriding the group_vars completely. The tag for package management is now singular "package" to match the other tags.

Note, the Jinja template for overrides has extra lines and spaces, this should be cleaned up but it parses properly, just doesn't lint cleanly, and I'm afraid of spending another 2-hours debugging spacing so I'm avoiding it to get improvements out now.