Improvements to the override template and tags
Hostname was tagged with override so removed that. Considered whether baseline is a useful tag due to overlap and it may be useful to apply some non-critical configuration so keeping it for the moment.

The override template has switched to human-based indexing (rather than 0, it starts with 1 now) to make the yaml easier to understand. After ruling out other syntax errors, the names changed from "server00_led", etc. to "ezcoo1_led1" matching the driver name given to reference the driver definition for ezcoo (similar for tesmart, etc.). This is much easier to read.

Getting rid of some of the space characters in empty lines has negative results on the rest of the yaml document so the override will complain (give warnings) about extra spaces at end-of-line, it did take too long to debug so I'm ignoring it at the moment as I risk breaking things for myself and others before spending holiday time with my parents.