Added `lstrip_blocks: true` to fix spaces in templates
I thought I'd be able to manage this with `{%-` but it's a pain with the number of unintended consequences. A friend was convinced nothing needed to be done with templates but I digress, without at least being implicit about whiteplace handling in both Ansible and Jinja, it's an mitigated mess. Thankfully, I didn't have to left-shift everything, just update the play.

With the change, the override.yaml now cleanly lints. I even caught an off-by-one error where a label or comment might have indicated a higher index number with `(loop.index + 1)` in a printf statement. That's been fixed as well.