Added schema for per-port names and wake on lan
This was a difficult change. I wanted the generated file to lint cleanly so it couldn't exceed 80 lines (this rule could be ignored but I'd prefer to keep things clean). There was also getting the driver, scheme, and lines to properly match. Another annoying issuee, while `ansible-lint` doesn't mind quoting, `ansible-lint --write` will remove quoting around the MAC addresses in the host_vars file which is not ideal, the colons break things.

Ultimately I'm happy to see it working, this is the first time I've set up WoL because it seemed too hard to want to deal with before but I've confirmed it's working as expected, though my list of which system is "host1" seems to be outdated so I'll need to correct my host_vars later.