Support for Ansible builder and content navigator.
This is a tough one to describe. There's a few extra steps for working with awx, Ansible Automation Platform, and ansible-navigator that needed to be addressed. Normally I would only use the CLI but someone mentioned they only use awx which implied an execution environment to me so I refreshed myself on how this worked and added configurations under `builder/`.

As a minor update, the overrides were separated into its own task file and `pikvm9.yml` was added to the `host_vars/` as I soldered the headers on a Raspberry Pi Zero 2 W to use for further validation. Minor changes were made to `ansible.cfg` to include galaxy configuration and `.gitignore` was updated to prevent the build context and ansible-navigator logs and runs from being committed by mistake. As a precaution, `galaxy-token` is also excluded should it be present.